+++ src/tasks/mod.rs
@@ -3,6 +3,7 @@ mod atom;
 mod dep;
 mod dotfiles;
 mod git;
+mod golang;
 mod hyper;
 mod jq;
 mod nodejs;
@@ -21,12 +22,12 @@ mod yq;
 pub fn sync() {
     // must be first
     dotfiles::sync();
-    #[cfg(windows)] windows::sync();
 
     alacritty::sync();
     atom::sync();
     dep::sync();
     git::sync();
+    golang::sync();
     hyper::sync();
     jq::sync();
     nodejs::sync();
@@ -38,18 +39,20 @@ pub fn sync() {
     tmux::sync();
     vim::sync();
     vscode::sync();
+    #[cfg(windows)]
+    windows::sync();
     yq::sync();
 }
 
 pub fn update() {
     // must be first
     dotfiles::update();
-    #[cfg(windows)] windows::update();
 
     alacritty::update();
     atom::update();
     dep::update();
     git::update();
+    golang::update();
     hyper::update();
     jq::update();
     nodejs::update();
@@ -61,5 +64,7 @@ pub fn update() {
     tmux::update();
     vim::update();
     vscode::update();
+    #[cfg(windows)]
+    windows::update();
     yq::update();
 }